Moved image code into IAGame function
Moved image code into IAGame function, added height and width to canvas.
Since there are only a certain set of images, I'm not making this
particularly extensible (outside loading of image data) and just
hard-coding the image URLs into the game, to save the trouble for such a
small app.