Fixed dots being out of position
I was accidentally pulling the dots toward their starting positions if
they were in the hand of the player. Oops. Fixed now. There's still a
bit of lag but I think that's due to some sort of input ordering. I'll
continue to work on that if I have time, but otherwise I think I'm happy
with that. Still have to add a reset button and make sure the two black
dots can't occupy the same spot.